fix: 地域コードとお天気コードを、別のファイルに切り離し
